requirements.txt
pandas
numpy
matplotlib
seaborn
pyarrow
scikit-learn
jupyter
pytest
sentence-transformers
langchain
langchain-text-splitters
faiss-cpu 
tqdm
transformers
torch
